requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-snowflake==5.6.0
apache-airflow-providers-amazon[s3fs]==8.19.0
apache-airflow-providers-slack==8.9.0
openai==1.17.1
langchain==0.2.6
tiktoken==0.7.0